more fixes

+++ EmagProject-master/src/main/java/finalproject/emag/service/DiscountService.java
@@ -8,25 +8,30 @@ import finalproject.emag.model.repository.DiscountRepository;
 import finalproject.emag.model.repository.ProductRepository;
 import finalproject.emag.util.DiscountNotify;
 import finalproject.emag.util.GetDate;
+import finalproject.emag.util.exception.BaseException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.Optional;
+import java.util.zip.DeflaterInputStream;
 
 
 @Service
-public final class DiscountService {
+public class DiscountService {
 
     @Autowired
     private DiscountRepository discountRepository;
     @Autowired
     private ProductRepository productRepository;
     @Autowired
+    private ProductService productService;
+    @Autowired
     private DiscountNotify discountNotify;
 
     @Transactional
-    public void addDiscount(PromotionProductDto product, Long productId) {
-        Product discounted = productRepository.findById(productId).get();
+    public void addDiscount(PromotionProductDto product, long productId) {
+        Product discounted = productService.checkIfProductExists(productId);
         double oldPrice = discounted.getPrice();
         Discount discount = Discount.builder().product(discounted).endDate(GetDate.getDate(product.getEndDate()))
                 .startDate(GetDate.getDate(product.getStartDate())).newPrice(product.getNewPrice())
@@ -40,11 +45,15 @@ public final class DiscountService {
     }
 
     @Transactional
-    public void removeDiscount(RemovePromotionDto promo) {
-        Product product = productRepository.findById(promo.getProductId()).get();
-        Discount discount  = discountRepository.findByProductId(promo.getProductId()).get();
-        product.setPrice(discount.getOldPrice());
-        productRepository.save(product);
-        discountRepository.delete(discount);
+    public void removeDiscount(long productId) {
+        Product product = productService.checkIfProductExists(productId);
+        Optional<Discount> savedDiscount  = discountRepository.findByProductId(productId);
+        if (savedDiscount.isPresent()) {
+            Discount discount = savedDiscount.get();
+            product.setPrice(discount.getOldPrice());
+            productRepository.save(product);
+            discountRepository.delete(discount);
+        }
+        else throw new BaseException("No such discount!");
     }
 }
